Refactor mocking code into new directory
